Revert "feat: copy fibre/ networking package from celestia-app-fibre (#6708)"
This reverts commit 60997a6b5d77d74a22e10fff185aafc925da66e2.
